{"schema":"decodex/commit/1","summary":"Split CLI command families","authority":"XY-1115"}
